+++ src/actions/modals/edit-modal/parts/main-edit-modal.tsx
@@ -5,6 +5,7 @@
  */
 import React, { FC, useCallback, useMemo, useState } from 'react';
 
+import styled from '@emotion/styled';
 import {
 	AnyColor,
 	Button,
@@ -36,7 +37,6 @@ import {
 } from '@zextras/carbonio-ui-commons';
 import { compact, find, includes, isEmpty, map } from 'lodash';
 import { useTranslation } from 'react-i18next';
-import styled from '@emotion/styled';
 
 import { GranteeChip } from './grantee-chip';
 import { ShareCalendarUrls } from './share-calendar-urls';
