+++ src/actions/modals/edit-modal/parts/svg-container.jsx
